django/requirements.txt
asgi-redis==1.0.0
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
certifi==2017.7.27.1
channels==1.0.2
chardet==3.0.4
click==6.7
constantly==15.1.0
daphne==1.3.0
Django==1.11
django-cors-middleware==1.3.1
django-filter==1.0.4
django-oauth-toolkit==1.0.0
djangorestframework==3.7.1
djangorestframework-simplejwt==3.0
freeze==1.0.10
graphene==2.0
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
iso8601==0.1.12
msgpack-python==0.4.8
mysqlclient==1.3.12
oauthlib==2.0.5
olefile==0.44
Pillow==4.3.0
promise==2.1
PyJWT==1.5.3
python-dotenv==0.7.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
txredisapi==1.4.4
typing==3.6.2
urllib3==1.22
zope.interface==4.4.3
